fix: prevent modal calls on selection
