+++ src/components/pages/home/HomeScreen.tsx
@@ -1,6 +1,6 @@
 import { useState } from 'react';
 
-import { Alert, Box, LinearProgress, Stack } from '@mui/material';
+import { Alert, Box, LinearProgress, Stack, useTheme } from '@mui/material';
 
 import { Button } from '@graasp/ui';
 
@@ -9,6 +9,7 @@ import {
   SelectionContextProvider,
   useSelectionContext,
 } from '@/components/main/list/SelectionContext';
+import { useDragSelection } from '@/components/main/list/useDragSelection';
 import { ITEM_PAGE_SIZE } from '@/config/constants';
 import { ShowOnlyMeChangeType } from '@/config/types';
 import { ItemLayoutMode, Ordering } from '@/enums';
@@ -43,6 +44,7 @@ const HomeScreenContent = ({ searchText }: { searchText: string }) => {
   const { data: currentMember } = hooks.useCurrentMember();
   const { itemTypes } = useFilterItemsContext();
   const [showOnlyMe, setShowOnlyMe] = useState(false);
+  const theme = useTheme();
 
   const { selectedIds, toggleSelection, clearSelection } =
     useSelectionContext();
@@ -65,6 +67,10 @@ const HomeScreenContent = ({ searchText }: { searchText: string }) => {
       // todo: adapt page size given the user window height
       { pageSize: ITEM_PAGE_SIZE },
     );
+
+  const DragSelection = useDragSelection({
+    adjustments: { marginTop: 100, marginLeft: theme.spacing(3) },
+  });
 
   const onShowOnlyMeChange: ShowOnlyMeChangeType = (checked) => {
     setShowOnlyMe(checked);
@@ -129,6 +135,7 @@ const HomeScreenContent = ({ searchText }: { searchText: string }) => {
 
     return (
       <>
+        <DragSelection />
         <Stack
           alignItems="space-between"
           direction="column"
